feat: refactor with examples folder
Instead of creating a whole new Rust project for every problem, create an example file.
Rust supports files in the examples directory as independent modules that can use dev-dependencies as well.
To support subdirs in the examples folder, an entry has to be made into the Cargo.lock giving the example a name and path to be ran with.